+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [vx.y.z] yyyy-mm-dd
+### Added
+- Added logger ability to log additional parameters.
+
+
 ## [v2.0.0] 2019-03-12
 ### Added
 - directory.contents() has a new `recursive` parameter.